+++ sdk/attestation/Azure.Security.Attestation/src/Models/PolicyModification.cs
@@ -1,16 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
-using System.Text;
-using Azure.Core;
+using Microsoft.TypeSpec.Generator.Customizations;
 
 namespace Azure.Security.Attestation
 {
     ///<summary>The result of a policy modification operation.</summary>
-    [CodeGenModel("PolicyModification")]
-    public partial struct PolicyModification
+    [CodeGenType("PolicyModification")]
+    public readonly partial struct PolicyModification
     {
     }
 }